requirements(1).txt
glfw==2.6.2
PyOpenGL==3.1.7
